refactor: move SaveToZeroCheck to own file
