+++ riscv_analysis/src/lints/mod.rs
@@ -30,3 +30,6 @@ pub use lost_callee_saved_register::*;
 
 mod stack;
 pub use stack::*;
+
+mod save_to_zero;
+pub use save_to_zero::*;
